refactor: split token and warp metadata surfaces
