+++ typescript/sdk/src/warp/WarpCore.ts
@@ -17,6 +17,7 @@ import {
 } from '@hyperlane-xyz/utils';
 import { Keypair } from '@solana/web3.js';
 
+import type { ChainMetadata } from '../metadata/chainMetadataTypes.js';
 import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../providers/ConfiguredMultiProtocolProvider.js';
 import { ProviderType } from '../providers/ProviderType.js';
 import {
@@ -26,7 +27,6 @@ import {
 import { IToken } from '../token/IToken.js';
 import { Token } from '../token/Token.js';
 import { TokenAmount } from '../token/TokenAmount.js';
-import { parseTokenConnectionId } from '../token/TokenConnection.js';
 import {
   LOCKBOX_STANDARDS,
   MINT_LIMITED_STANDARDS,
@@ -65,6 +65,7 @@ import {
   WarpTxCategory,
   WarpTypedTransaction,
 } from './types.js';
+import { WarpRouteGraph, buildWarpRouteTokens } from './WarpRouteGraph.js';
 
 export interface WarpCoreOptions {
   logger?: Logger;
@@ -73,9 +74,7 @@ export interface WarpCoreOptions {
   routeBlacklist?: RouteBlacklist;
 }
 
-export class WarpCore {
-  public readonly multiProvider: MultiProtocolProvider<{ mailbox?: Address }>;
-  public readonly tokens: Token[];
+export class WarpCore extends WarpRouteGraph<Token> {
   public readonly localFeeConstants: FeeConstantConfig;
   public readonly interchainFeeConstants: FeeConstantConfig;
   public readonly routeBlacklist: RouteBlacklist;
@@ -86,8 +85,7 @@ export class WarpCore {
     tokens: Token[],
     options?: WarpCoreOptions,
   ) {
-    this.multiProvider = multiProvider;
-    this.tokens = tokens;
+    super(multiProvider, tokens);
     this.localFeeConstants = options?.localFeeConstants || [];
     this.interchainFeeConstants = options?.interchainFeeConstants || [];
     this.routeBlacklist = options?.routeBlacklist || [];
@@ -107,44 +105,21 @@ export class WarpCore {
     multiProvider: MultiProtocolProvider<{ mailbox?: Address }>,
     config: unknown,
   ): WarpCore {
-    // Validate and parse config data
     const parsedConfig = WarpCoreConfigSchema.parse(config);
-    // Instantiate all tokens
-    const tokens = parsedConfig.tokens.map(
-      (t) =>
+    const tokens = buildWarpRouteTokens(
+      parsedConfig,
+      (token) =>
         new Token({
-          ...t,
-          addressOrDenom: t.addressOrDenom || '',
+          ...token,
+          addressOrDenom: token.addressOrDenom || '',
           connections: undefined,
         }),
     );
-    // Connect tokens together
-    parsedConfig.tokens.forEach((config, i) => {
-      for (const connection of config.connections || []) {
-        const token1 = tokens[i];
-        const { chainName, addressOrDenom } = parseTokenConnectionId(
-          connection.token,
-        );
-        // If token1 has a warpRouteId, token2 must share it — disambiguates
-        // tokens at the same address (e.g. M0 Portal: mUSD, wM, USDSC)
-        const token2 = tokens.find(
-          (t) =>
-            t.chainName === chainName &&
-            t.addressOrDenom === addressOrDenom &&
-            (!token1.warpRouteId || t.warpRouteId === token1.warpRouteId),
-        );
-        assert(
-          token2,
-          `Connected token not found: ${chainName} ${addressOrDenom}`,
-        );
-        token1.addConnection({
-          ...connection,
-          token: token2,
-        });
-      }
-    });
-    // Create new Warp
     return new WarpCore(multiProvider, tokens, parsedConfig.options);
+  }
+
+  protected override createNativeToken(chainMetadata: ChainMetadata): Token {
+    return Token.FromChainMetadataNativeToken(chainMetadata);
   }
 
   /**
@@ -159,12 +134,15 @@ export class WarpCore {
     recipient,
     destinationToken,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     sender?: Address;
     recipient: Address;
     destinationToken?: IToken;
-  }): Promise<{ igpQuote: TokenAmount; tokenFeeQuote?: TokenAmount }> {
+  }): Promise<{
+    igpQuote: TokenAmount<IToken>;
+    tokenFeeQuote?: TokenAmount<IToken>;
+  }> {
     this.logger.debug(`Fetching interchain transfer quote to ${destination}`);
     const { amount, token: originToken } = originTokenAmount;
     const originName = originToken.chainName;
@@ -241,7 +219,7 @@ export class WarpCore {
       igpToken = searchResult;
     }
 
-    let feeTokenAmount: TokenAmount | undefined;
+    let feeTokenAmount: TokenAmount<IToken> | undefined;
     if (feeAmount) {
       // empty address or zero address is native route
       if (!feeTokenAddress || isZeroishAddress(feeTokenAddress)) {
@@ -280,8 +258,8 @@ export class WarpCore {
     destination: ChainNameOrId;
     sender: Address;
     senderPubKey?: HexString;
-    interchainFee?: TokenAmount;
-    tokenFeeQuote?: TokenAmount;
+    interchainFee?: TokenAmount<IToken>;
+    tokenFeeQuote?: TokenAmount<IToken>;
     destinationToken?: IToken;
   }): Promise<TransactionFeeEstimate> {
     this.logger.debug(`Estimating local transfer gas to ${destination}`);
@@ -386,10 +364,10 @@ export class WarpCore {
     destination: ChainNameOrId;
     sender: Address;
     senderPubKey?: HexString;
-    interchainFee?: TokenAmount;
-    tokenFeeQuote?: TokenAmount;
+    interchainFee?: TokenAmount<IToken>;
+    tokenFeeQuote?: TokenAmount<IToken>;
     destinationToken?: IToken;
-  }): Promise<TokenAmount> {
+  }): Promise<TokenAmount<IToken>> {
     const originMetadata = this.multiProvider.getChainMetadata(
       originToken.chainName,
     );
@@ -431,12 +409,12 @@ export class WarpCore {
     destinationToken,
     quotedCalls,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     sender: Address;
     recipient: Address;
-    interchainFee?: TokenAmount;
-    tokenFeeQuote?: TokenAmount;
+    interchainFee?: TokenAmount<IToken>;
+    tokenFeeQuote?: TokenAmount<IToken>;
     destinationToken?: IToken;
     /** When provided, builds an atomic QuotedCalls.execute() tx instead of separate approve+transfer */
     quotedCalls?: QuotedCallsParams;
@@ -640,7 +618,7 @@ export class WarpCore {
     recipient,
     destinationToken,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     sender: Address;
     recipient: Address;
@@ -1026,7 +1004,7 @@ export class WarpCore {
     senderPubKey,
     destinationToken,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     recipient: Address;
     sender: Address;
@@ -1086,7 +1064,7 @@ export class WarpCore {
     sender,
     senderPubKey,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     destinationToken: IToken;
     recipient: Address;
@@ -1139,14 +1117,14 @@ export class WarpCore {
     feeEstimate,
     destinationToken,
   }: {
-    balance: TokenAmount;
+    balance: TokenAmount<IToken>;
     destination: ChainNameOrId;
     recipient: Address;
     sender: Address;
     senderPubKey?: HexString;
     feeEstimate?: WarpCoreFeeEstimate;
     destinationToken?: IToken;
-  }): Promise<TokenAmount> {
+  }): Promise<TokenAmount<IToken>> {
     const originToken = balance.token;
 
     if (!feeEstimate) {
@@ -1210,7 +1188,7 @@ export class WarpCore {
     destination,
     destinationToken,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     destinationToken?: IToken;
   }): Promise<boolean> {
@@ -1282,7 +1260,7 @@ export class WarpCore {
     originTokenAmount,
     owner,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     owner: Address;
   }): Promise<boolean> {
     const { token, amount } = originTokenAmount;
@@ -1311,7 +1289,7 @@ export class WarpCore {
     senderPubKey,
     destinationToken,
   }: {
-    originTokenAmount: TokenAmount;
+    originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     recipient: Address;
     sender: Address;
@@ -1444,7 +1422,7 @@ export class WarpCore {
    * Ensure token amount is valid
    */
   protected async validateAmount(
-    originTokenAmount: TokenAmount,
+    originTokenAmount: TokenAmount<IToken>,
     destination: ChainNameOrId,
     recipient: Address,
     destinationToken?: IToken,
@@ -1495,7 +1473,7 @@ export class WarpCore {
    * Ensure the sender has sufficient balances for transfer and interchain gas
    */
   protected async validateTokenBalances(
-    originTokenAmount: TokenAmount,
+    originTokenAmount: TokenAmount<IToken>,
     destination: ChainNameOrId,
     sender: Address,
     recipient: Address,
@@ -1581,7 +1559,7 @@ export class WarpCore {
    * Ensure the sender has sufficient balances for transfer and interchain gas
    */
   protected async validateDestinationCollateral(
-    originTokenAmount: TokenAmount,
+    originTokenAmount: TokenAmount<IToken>,
     destination: ChainNameOrId,
     destinationToken?: IToken,
   ): Promise<Record<string, string> | null> {
@@ -1601,7 +1579,7 @@ export class WarpCore {
    * Ensure the sender has sufficient balances for minting
    */
   protected async validateDestinationRateLimit(
-    originTokenAmount: TokenAmount,
+    originTokenAmount: TokenAmount<IToken>,
     destination: ChainNameOrId,
     destinationToken?: IToken,
   ): Promise<Record<string, string> | null> {
@@ -1675,7 +1653,7 @@ export class WarpCore {
    * Ensure the sender has sufficient balances for transfer and interchain gas
    */
   protected async validateOriginCollateral(
-    originTokenAmount: TokenAmount,
+    originTokenAmount: TokenAmount<IToken>,
   ): Promise<Record<string, string> | null> {
     const adapter = originTokenAmount.token.getAdapter(this.multiProvider);
 
@@ -1692,104 +1670,5 @@ export class WarpCore {
     }
 
     return null;
-  }
-
-  protected resolveDestinationToken({
-    originToken,
-    destination,
-    destinationToken,
-  }: {
-    originToken: IToken;
-    destination: ChainNameOrId;
-    destinationToken?: IToken;
-  }): IToken {
-    const destinationName = this.multiProvider.getChainName(destination);
-    const destinationCandidates = originToken
-      .getConnections()
-      .filter((connection) => connection.token.chainName === destinationName)
-      .map((connection) => connection.token);
-
-    assert(
-      destinationCandidates.length > 0,
-      `No connection found for ${destinationName}`,
-    );
-
-    if (destinationToken) {
-      assert(
-        destinationToken.chainName === destinationName,
-        `Destination token chain mismatch for ${destinationName}`,
-      );
-      const matchedToken = destinationCandidates.find(
-        (candidate) =>
-          candidate.equals(destinationToken) ||
-          candidate.addressOrDenom.toLowerCase() ===
-            destinationToken.addressOrDenom.toLowerCase(),
-      );
-      assert(
-        matchedToken,
-        `Destination token ${destinationToken.addressOrDenom} is not connected from ${originToken.chainName} to ${destinationName}`,
-      );
-      return matchedToken;
-    }
-
-    assert(
-      destinationCandidates.length === 1,
-      `Ambiguous route to ${destinationName}; specify destination token`,
-    );
-    return destinationCandidates[0];
-  }
-
-  /**
-   * Search through token list to find token with matching chain and address
-   */
-  findToken(
-    chainName: ChainName,
-    addressOrDenom?: Address | string,
-  ): Token | null {
-    if (!addressOrDenom) return null;
-
-    const results = this.tokens.filter(
-      (token) =>
-        token.chainName === chainName &&
-        token.addressOrDenom.toLowerCase() === addressOrDenom.toLowerCase(),
-    );
-
-    if (results.length === 1) return results[0];
-
-    if (results.length > 1)
-      throw new Error(`Ambiguous token search results for ${addressOrDenom}`);
-
-    // If the token is not found, check to see if it matches the denom of chain's native token
-    // This is a convenience so WarpConfigs don't need to include definitions for native tokens
-    const chainMetadata = this.multiProvider.getChainMetadata(chainName);
-    if (chainMetadata.nativeToken?.denom === addressOrDenom) {
-      return Token.FromChainMetadataNativeToken(chainMetadata);
-    }
-
-    return null;
-  }
-
-  /**
-   * Get the list of chains referenced by the tokens in this WarpCore
-   */
-  getTokenChains(): ChainName[] {
-    return [...new Set(this.tokens.map((t) => t.chainName)).values()];
-  }
-
-  /**
-   * Get the subset of tokens whose chain matches the given chainName
-   */
-  getTokensForChain(chainName: ChainName): Token[] {
-    return this.tokens.filter((t) => t.chainName === chainName);
-  }
-
-  /**
-   * Get the subset of tokens whose chain matches the given chainName
-   * and which are connected to a token on the given destination chain
-   */
-  getTokensForRoute(origin: ChainName, destination: ChainName): Token[] {
-    return this.tokens.filter(
-      (t) => t.chainName === origin && t.getConnectionForChain(destination),
-    );
   }
 }
